Cleans up merge conflicts and polish
I started work on this a month ago and then put it on the backburner when I worked on other projects. I'm coming back to it now.

This PR just redoes a lot of area-based nomenclature. Although (9abb9b8bdbb48ba38acc420845961d999f037f8a) was great and getting rid of the obtuse "secondary" thing, there were still a lot of places I didn't address. I decided to re-draw a lot of maintenance sections, but this time instead of focusing on getting rid of weildy areas, I focused on getting good descriptive names and patterns.

For example, on IceBoxStation, I made it such that you get your normal station-naming schemes on the top floor, and you only get departmental maintenance on the bottom floor. I've also split out a lot of rooms that were in these bulky maintenance sections and split them into their own functioning rooms. I've largely forgotten a lot of the changes, but those were the two main areas I was looking at as I redrew all of this.