Icemoon Ruin: The Abandoned Shipping Yard
Hey there,

We recently got some nice stuff in regards to clocks, flags, and shipping containers- so I decided to make something very, very simple to give them a small home in game. I eagerly anticipate their usage in more and more maps, but I really have been thinking about this sorta small idea for a good while now, and I think this is a nice addition.

Not everything needs to be chock-full of lore and detail either, you know? Sometimes an abandoned shipyard full of old containers is just that.