datumization
I spoke to Willard to figure out a better way to get a list of potential bubbles we could swap between, and we settled on making datums be cool. I wish there was a way to do it without Newing every datum, but I'm not sure if there is one. Looks cool though!