Culls MAKE_AI from the View Variables Dropdown
For some reason, this relied on some snowflake behavior in which you'd create a list with the player in it by calling Topic() in some really odd way that I didn't really see anywhere else. Just odd overall, and as Shiz recommended in the linked report- most people use the Player Panel for this instead (and I don't see the intrinsic value in this being redundant).