Running Integration Tests requires Linters/Compile/Windows Build to Pass
Hey there,

This is just a small stop-guard I came up with. Recently, there's been an elevated amount of CI "clogging", which results in auto-merging of PRs only occurring roughly two hours after the auto-merge button has been hit due to an excessive number of alloted GitHub runners taking up space.

One thing that I recently noted is that even though linters were failing or that the windows build was failing, we would still allocate runners to do the exact same thing! This means that even if one runner reported that it failed to build, we would still take up seven more runners and fail out at the compilation step over-and-over again. So, this PR just just prevents that excess chew.

These three workflows are quite resistant to failure (and are required to pass for merging anyways), so let's just ensure we don't keep driving a car into a wall if we know it's not going to smash through.